add sqs
