+++ lib/database-stack/database-stack.ts
@@ -6,9 +6,11 @@ import * as lambda from "aws-cdk-lib/aws-lambda";
 import * as cdk from 'aws-cdk-lib';
 import {  join } from 'path';
 import * as apigateway from "aws-cdk-lib/aws-apigateway";
+import * as sqs from "aws-cdk-lib/aws-sqs";
+import { SqsEventSource } from "aws-cdk-lib/aws-lambda-event-sources";
 
 const dotenv = require("dotenv");
-dotenv.config(); 
+dotenv.config();
 
 const PRODUCTS_TABLE_NAME = process.env.PRODUCTS_TABLE_NAME;
 const STOCK_TABLE_NAME = process.env.STOCK_TABLE_NAME;
@@ -16,13 +18,14 @@ const REGION = process.env.AWS_REGION;
 const LAMBDA_API_URL = process.env.LAMBDA_API_URL;
 
 if (!PRODUCTS_TABLE_NAME || !STOCK_TABLE_NAME || !REGION || !LAMBDA_API_URL) {
-  throw new Error("Missing required environment variables. Please check your .env file.");
+  throw new Error(
+    "Missing required environment variables. Please check your .env file."
+  );
 }
 
 export class DatabaseStack extends Stack {
   constructor(scope: Construct, id: string, props?: StackProps) {
     super(scope, id, props);
-
 
     const productsTable = new dynamodb.Table(this, "Products", {
       tableName: PRODUCTS_TABLE_NAME,
@@ -36,34 +39,63 @@ export class DatabaseStack extends Stack {
     const stockTable = new dynamodb.Table(this, "StockTable", {
       tableName: STOCK_TABLE_NAME,
       partitionKey: {
-        name: "product_id", 
+        name: "product_id",
         type: dynamodb.AttributeType.STRING,
       },
       removalPolicy: cdk.RemovalPolicy.DESTROY,
     });
 
-    const createProductLambda = new lambda.Function(this, 'lambda-function', {
+    const catalogItemsQueue = new sqs.Queue(this, "CatalogItemsQueue", {
+      queueName: "catalogItemsQueue",
+      visibilityTimeout: cdk.Duration.seconds(30),
+    });
+
+    const catalogBatchProcess = new lambda.Function(
+      this,
+      "CatalogBatchProcess",
+      {
+        runtime: lambda.Runtime.NODEJS_20_X,
+        handler: "catalog-batch-process.catalogBatchProcess",
+        code: lambda.Code.fromAsset(join(__dirname, "./lambdas")),
+        environment: {
+          PRODUCTS_TABLE_NAME: productsTable.tableName,
+          STOCK_TABLE_NAME: stockTable.tableName,
+        },
+      }
+    );
+
+    catalogBatchProcess.addEventSource(
+      new SqsEventSource(catalogItemsQueue, {
+        batchSize: 5,
+      })
+    );
+
+    const createProductLambda = new lambda.Function(this, "lambda-function", {
       runtime: lambda.Runtime.NODEJS_20_X,
       memorySize: 1024,
       timeout: cdk.Duration.seconds(5),
-      handler: 'handler.createProduct',
-      code: lambda.Code.fromAsset(join(__dirname, './')),
-      environment: {
-        PRODUCTS_TABLE_NAME: productsTable.tableName,
-        STOCK_TABLE_NAME: stockTable.tableName,     
-      }
-    });
-    const getProductsListLambda = new lambda.Function(this, "GetProductsListFunction", {
-      runtime: lambda.Runtime.NODEJS_20_X,
-      memorySize: 512,
-      timeout: cdk.Duration.seconds(5),
-      handler: "handler.getProductsList",
-      code: lambda.Code.fromAsset(join(__dirname, "./")),
+      handler: "create-product.createProduct",
+      code: lambda.Code.fromAsset(join(__dirname, "./lambdas")),
       environment: {
         PRODUCTS_TABLE_NAME: productsTable.tableName,
         STOCK_TABLE_NAME: stockTable.tableName,
       },
     });
+    const getProductsListLambda = new lambda.Function(
+      this,
+      "GetProductsListFunction",
+      {
+        runtime: lambda.Runtime.NODEJS_20_X,
+        memorySize: 512,
+        timeout: cdk.Duration.seconds(5),
+        handler: "get-products-list.getProductsList",
+        code: lambda.Code.fromAsset(join(__dirname, "./lambdas")),
+        environment: {
+          PRODUCTS_TABLE_NAME: productsTable.tableName,
+          STOCK_TABLE_NAME: stockTable.tableName,
+        },
+      }
+    );
 
     const getProductByIdLambda = new lambda.Function(
       this,
@@ -72,8 +104,8 @@ export class DatabaseStack extends Stack {
         runtime: lambda.Runtime.NODEJS_20_X,
         memorySize: 512,
         timeout: cdk.Duration.seconds(5),
-        handler: "handler.getProductById",
-        code: lambda.Code.fromAsset(join(__dirname, "./")),
+        handler: "get-product-by-id.getProductById",
+        code: lambda.Code.fromAsset(join(__dirname, "./lambdas")),
         environment: {
           PRODUCTS_TABLE_NAME: productsTable.tableName,
           STOCK_TABLE_NAME: stockTable.tableName,
@@ -85,15 +117,17 @@ export class DatabaseStack extends Stack {
       runtime: lambda.Runtime.NODEJS_20_X,
       memorySize: 512,
       timeout: cdk.Duration.seconds(5),
-      handler: "handler.getStock",
-      code: lambda.Code.fromAsset(join(__dirname, "./")),
+      handler: "get-stock.getStock",
+      code: lambda.Code.fromAsset(join(__dirname, "./lambdas")),
       environment: {
         STOCK_TABLE_NAME: stockTable.tableName,
       },
     });
 
     productsTable.grantWriteData(createProductLambda);
+    productsTable.grantWriteData(catalogBatchProcess);
     stockTable.grantWriteData(createProductLambda);
+    stockTable.grantWriteData(catalogBatchProcess);
     productsTable.grantReadData(getProductsListLambda);
     productsTable.grantReadData(getStockLambda);
     stockTable.grantReadData(getStockLambda);
@@ -105,18 +139,27 @@ export class DatabaseStack extends Stack {
       description: "API for managing products and stock.",
     });
 
-    const createProductIntegration = new apigateway.LambdaIntegration(createProductLambda);
-    const getProductsListIntegration = new apigateway.LambdaIntegration(getProductsListLambda);
-    const getProductByIdIntegration = new apigateway.LambdaIntegration(getProductByIdLambda);
+    const createProductIntegration = new apigateway.LambdaIntegration(
+      createProductLambda
+    );
+    const getProductsListIntegration = new apigateway.LambdaIntegration(
+      getProductsListLambda
+    );
+    const getProductByIdIntegration = new apigateway.LambdaIntegration(
+      getProductByIdLambda
+    );
     productsTable.grantReadData(getProductByIdLambda);
 
-    const productsResource = api.root.addResource("products"); 
-    productsResource.addMethod("POST", createProductIntegration); 
-    productsResource.addMethod("GET", getProductsListIntegration); 
-    productsResource.addResource("{productId}").addMethod("GET", getProductByIdIntegration);
+    const productsResource = api.root.addResource("products");
+    productsResource.addMethod("POST", createProductIntegration);
+    productsResource.addMethod("GET", getProductsListIntegration);
+    productsResource
+      .addResource("{productId}")
+      .addMethod("GET", getProductByIdIntegration);
 
-    const getStockIntegration = new apigateway.LambdaIntegration(getStockLambda);
+    const getStockIntegration = new apigateway.LambdaIntegration(
+      getStockLambda
+    );
     api.root.addResource("stock").addMethod("GET", getStockIntegration);
-  
   }
 }
